+++ etherparse/src/transport/tcp_header.rs
@@ -437,7 +437,7 @@ impl TcpHeader {
         // SAFETY: Safe as the header len can not exceed the maximum length
         // of the header.
         unsafe {
-            result.set_len(self.header_len() as usize);
+            result.set_len(self.header_len());
         }
 
         result
@@ -460,7 +460,7 @@ impl TcpHeader {
         payload: &[u8],
     ) -> Result<u16, ValueTooBigError<usize>> {
         // check that the total length fits into the tcp length field
-        let max_payload = usize::from(core::u16::MAX) - usize::from(self.header_len());
+        let max_payload = usize::from(core::u16::MAX) - self.header_len();
         if max_payload < payload.len() {
             return Err(ValueTooBigError {
                 actual: payload.len(),
@@ -498,7 +498,7 @@ impl TcpHeader {
         payload: &[u8],
     ) -> Result<u16, ValueTooBigError<usize>> {
         // check that the total length fits into the tcp length field
-        let max_payload = (core::u32::MAX as usize) - usize::from(self.header_len());
+        let max_payload = (core::u32::MAX as usize) - self.header_len();
         if max_payload < payload.len() {
             return Err(ValueTooBigError {
                 actual: payload.len(),
